feat: add support for logging to file with configurable options
The changes in this commit add support for logging to a file in addition to the standard output. The logging options can be configured through environment variables. The following environment variables are used:

- `LOG_TO_FILE`: Set to "true" to enable logging to file
- `LOG_FILENAME`: The path and filename of the log file
- `LOG_MAX_SIZE`: The maximum size of the log file in megabytes
- `LOG_MAX_AGE`: The maximum age of log files in days
- `LOG_MAX_BACKUPS`: The maximum number of log file backups to keep
- `LOG_LEVEL`: The log level to use (debug, info, warn, error)

If `LOG_TO_FILE` is set to "true", the log messages will be written to the specified file using the JSON format. Otherwise, the log messages will be written to the standard output.

The `initLogger` function is introduced to handle the initialization